test(update): consolidate npm owner coverage within shard budget
Keep custom prefix and lifecycle coverage in the existing npm global suite without increasing the infra test-root budget or dropping assertions.
